requirements.txt
praw==7.7.1
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.18
sentence-transformers==2.2.2
tqdm==4.66.1
numpy==1.24.3
Flask
Flask-Cors 